Add files via upload
This code was a project of mine that I created while studying at the academy. It performs the function of a "task book", where you can edit, add, search and clear tasks. 
The code is based on the file system and the fsrteam module. We recommend uploading files in the following format before launching.txt files to your computer, otherwise they will be created automatically.